requirements.txt
--extra-index-url https://download.pytorch.org/whl/rocm5.7
torch
torchvision
torchmetrics[image]
tensorboard
lightning
lmdb
zstandard
tqdm
matplotlib
numpy
rich
scikit-learn
gitpython
scipy
opencv-python
pandas
requests
dacite
jaxtyping
beartype